# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pushboy API
  version: 0.1.0
  summary: Self-hosted APNS, FCM, and Live Activity delivery API.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  description: |
    Pushboy exposes a small HTTP API for users, device tokens, topics,
    notification jobs, and Live Activity dispatches. The current API does not
    include built-in authentication; deploy it behind a trusted network,
    gateway, or reverse proxy until first-party API auth is added.
servers:
  - url: http://localhost:8080
    description: Local development
...
security: []
paths:
...
  /v1/users:
    get:
      tags: [Users]
      summary: List users
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Users ordered by creation time descending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
        "400":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/:
    post:
      tags: [Users]
      summary: Create a user
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/TextError"
        "409":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/{userID}:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      tags: [Users]
      summary: Get a user
      responses:
        "200":
          description: User found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
    delete:
      tags: [Users]
      summary: Delete a user
      responses:
        "204":
          description: User deleted.
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/tokens:
    post:
      tags: [Tokens]
      summary: Register a push token
      description: Creates the user if the provided id does not exist. Generates a user id if id is omitted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTokenRequest"
      responses:
        "201":
          description: Token registered.
          content:
            application/json:
              schema:
                type: object
                required: [user, token]
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  token:
                    $ref: "#/components/schemas/Token"
        "400":
          $ref: "#/components/responses/TextError"
        "409":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/{userID}/tokens:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      tags: [Tokens]
      summary: List a user's push tokens
      responses:
        "200":
          description: Tokens for the user. Unknown users currently return an empty list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/{userID}/tokens/{tokenID}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/TokenID"
    delete:
      tags: [Tokens]
      summary: Delete a push token
      description: The current handler deletes by tokenID; userID is part of the route but is not used for authorization.
      responses:
        "204":
          description: Token deleted.
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/{userID}/subscriptions/{topicID}:
    parameters:
      - $ref: "#/components/parameters/UserID"
      - $ref: "#/components/parameters/TopicID"
    post:
      tags: [Subscriptions]
      summary: Subscribe a user to a topic
      responses:
        "201":
          description: Subscription created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserTopicSubscription"
        "404":
          $ref: "#/components/responses/TextError"
        "409":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
    delete:
      tags: [Subscriptions]
      summary: Unsubscribe a user from a topic
      responses:
        "204":
          description: Subscription deleted.
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/{userID}/subscriptions:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      tags: [Subscriptions]
      summary: List a user's topic subscriptions
      responses:
        "200":
          description: Subscriptions for the user. Unknown users currently return an empty list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserTopicSubscription"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/{userID}/notifications:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      tags: [Notifications]
      summary: List notifications eligible for a user
      description: Returns direct user sends and topic sends where the user is currently subscribed and the job effective time is on or after the subscription time. This is eligibility, not delivery history.
      parameters:
        - $ref: "#/components/parameters/NotificationStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Eligible notification jobs ordered by effective time descending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPublishJobs"
        "400":
          $ref: "#/components/responses/TextError"
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/users/{userID}/send:
    parameters:
      - $ref: "#/components/parameters/UserID"
    post:
      tags: [Notifications]
      summary: Send or schedule a notification to one user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationRequest"
      responses:
        "202":
          description: Job queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueuedJobResponse"
        "400":
          $ref: "#/components/responses/TextError"
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
        "503":
          $ref: "#/components/responses/TextError"
  /v1/topics/:
    get:
      tags: [Topics]
      summary: List topics
      responses:
        "200":
          description: Topics.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Topic"
        "500":
          $ref: "#/components/responses/TextError"
    post:
      tags: [Topics]
      summary: Create a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTopicRequest"
      responses:
        "201":
          description: Topic created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
        "400":
          $ref: "#/components/responses/TextError"
        "409":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/topics/{topicID}:
    parameters:
      - $ref: "#/components/parameters/TopicID"
    get:
      tags: [Topics]
      summary: Get a topic
      responses:
        "200":
          description: Topic found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topic"
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
    delete:
      tags: [Topics]
      summary: Delete a topic
      responses:
        "204":
          description: Topic deleted.
        "404":
          $ref: "#/components/responses/TextError"
        "409":
          description: Topic is still referenced by a Live Activity channel.
          content:
            text/plain:
              schema:
                type: string
        "500":
          $ref: "#/components/responses/TextError"
  /v1/topics/{topicID}/count:
    parameters:
      - $ref: "#/components/parameters/TopicID"
    get:
      tags: [Topics]
      summary: Count topic subscribers
      responses:
        "200":
          description: Subscriber count. Missing topics currently can return 0.
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count:
                    type: integer
                    example: 42
        "500":
          $ref: "#/components/responses/TextError"
  /v1/topics/{topicID}/subscribers:
    parameters:
      - $ref: "#/components/parameters/TopicID"
    get:
      tags: [Subscriptions]
      summary: List topic subscribers
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Current subscribers ordered by subscription time descending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTopicSubscribers"
        "400":
          $ref: "#/components/responses/TextError"
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
  /v1/topics/{topicID}/notifications:
    parameters:
      - $ref: "#/components/parameters/TopicID"
    get:
      tags: [Notifications]
      summary: List notifications for a topic
      parameters:
        - $ref: "#/components/parameters/NotificationStatus"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Topic notification jobs ordered by effective time descending.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPublishJobs"
        "400":
          $ref: "#/components/responses/TextError"
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
...
  /v1/notifications/{jobID}:
    parameters:
      - $ref: "#/components/parameters/JobID"
    get:
      tags: [Notifications]
      summary: Get notification job
      responses:
        "200":
          description: Notification job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishJob"
        "404":
          $ref: "#/components/responses/TextError"
        "500":
          $ref: "#/components/responses/TextError"
...
components:
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque URL-safe cursor returned by the previous page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Defaults to 50 and must be between 1 and 200.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    TextError:
      description: Plain text error response.
      content:
        text/plain:
          schema:
            type: string
            example: Bad request
  schemas:
...
    CreateUserRequest:
      type: object
      properties:
        id:
          type: string
          description: Optional caller-provided user id. A UUID is generated when omitted.
          example: user-123
...
    PaginatedUsers:
      type: object
      required: [items, next_cursor, has_more]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        next_cursor:
          type: string
          example: ""
        has_more:
          type: boolean
          example: false
...